feat: update endpoints to return gas prices and approval costs
