+++ apps/axelarscan-amplifier-apis/src/configs/feesAndPrices.ts
@@ -1,0 +1,10 @@
+export default [
+  {
+    name: "Avalanche",
+    amplifierChainId: "avalanche",
+    description: null,
+    gasPriceGwei: 28,
+    approvalCost: 425000,
+    updated: "2024-05-29",
+  },
+];
