+++ apps/axelarscan-amplifier-apis/src/index.ts
@@ -1,25 +1,27 @@
 import { OpenAPIRouter } from "@cloudflare/itty-router-openapi";
-import { TaskFetch } from "./endpoints/taskFetch";
-import { ChainList } from "./endpoints/chainList";
+
+import { ChainList } from "./endpoints/gasPricesAllChains";
+import { ChainFetch } from "./endpoints/getPricesForChain";
 
 export const router = OpenAPIRouter({
-	docs_url: "/",
+  docs_url: "/",
 });
 
 router.get("/api/gasPrices/", ChainList);
-router.get("/api/gasPrices/:chainSlug/", TaskFetch);
+router.get("/api/gasPrices/:amplifierChainId/", ChainFetch);
 
 // 404 for everything else
 router.all("*", () =>
-	Response.json(
-		{
-			success: false,
-			error: "Route not found",
-		},
-		{ status: 404 }
-	)
+  Response.json(
+    {
+      success: false,
+      error: "Route not found",
+    },
+    { status: 404 }
+  )
 );
 
+// @ts-ignore
 export default {
-	fetch: router.handle,
+  fetch: router.handle,
 };
